Mobile - Refactor of the Social Media component and fix some issues with SCSS rules and title sizes of the section
