+++ src/components/shared/icons/CvIcon.vue
