+++ src/components/shared/icons/InstagramIcon.vue
