+++ src/components/shared/icons/RightArrowIcon.vue
