+++ src/styles/_base.scss
@@ -14,8 +14,8 @@
 }
 
 body {
-  @include flex-center;
   align-items: inherit;
   margin: 30px;
   min-height: calc(100vh - 60px);
+  @include flex-center;
 }
