+++ src/styles/components/_articles.scss
@@ -3,7 +3,6 @@
 
 .articles {
   &-title {
-    @include responsive-font(title);
     color: $neutral-100;
     font-family: $font-family-secondary;
     font-weight: 800;
@@ -11,7 +10,7 @@
     margin-bottom: 45px;
     margin-top: 60px;
     @include flex-center;
-    @include responsive-font(subheading);
+    @include responsive-font(title);
   }
 
   &-link {
