+++ src/styles/main.scss
@@ -4,6 +4,7 @@
 
 // Components styles
 @use './components/cover-letter' as *;
+@use './components/shared/socialmedia' as *;
 @use './components/about' as *;
 @use './components/projects' as *;
 @use './components/shared/projectcard' as *;
